requirements.txt
python-telegram-bot[job-queue,rate-limiter]>=20.0
python-dotenv>=1.0
